🚚 api: move csv for certified clea to results context
