+++ api/src/certification/results/infrastructure/utils/csv/certification-results/get-clea-certified-candidate-csv.js
@@ -1,6 +1,6 @@
 import dayjs from 'dayjs';
 
-import { getCsvContent } from '../write-csv-utils.js';
+import { getCsvContent } from '../../../../../../shared/infrastructure/utils/csv/write-csv-utils.js';
 
 const getCleaCertifiedCandidateCsv = async function ({ cleaCertifiedCandidates }) {
   const fileHeaders = _buildFileHeadersForCleaCandidates(cleaCertifiedCandidates);
