+++ api/tests/certification/results/integration/infrastructure/utils/csv/certification-results/get-clea-certified-candidate-csv_test.js
@@ -1,5 +1,5 @@
-import { getCleaCertifiedCandidateCsv } from '../../../../../../lib/infrastructure/utils/csv/certification-results/get-clea-certified-candidate-csv.js';
-import { domainBuilder, expect } from '../../../../../test-helper.js';
+import { getCleaCertifiedCandidateCsv } from '../../../../../../../../src/certification/results/infrastructure/utils/csv/certification-results/get-clea-certified-candidate-csv.js';
+import { domainBuilder, expect } from '../../../../../../../test-helper.js';
 
 describe('Integration | Application | UseCases | certification-results | get-clea-certified-candidate-csv', function () {
   context('#getCleaCertifiedCandidateCsv', function () {
